Move BER Functional Test into Unit Test dir
Test indiv Prefect Task functions in Unit & leave
functional test for running pipelines from user
perspective or for running indiv Tasks if user story
is just to extract data from source via `drem`
