+++ tests/unit/transform/test_ber.py
@@ -1,9 +1,17 @@
 import pandas as pd
 
 from pandas.testing import assert_frame_equal
+from tdda.referencetest.referencetest import ReferenceTest
 
+import drem
+
+from drem.filepaths import UTEST_DATA_TRANSFORM
 from drem.transform.ber import _bin_year_of_construction_as_in_census
 from drem.transform.ber import _extract_dublin_rows
+
+
+BER_RAW = UTEST_DATA_TRANSFORM / "ber_raw.parquet"
+BER_CLEAN = UTEST_DATA_TRANSFORM / "ber_clean.csv"
 
 
 def test_extract_dublin_rows() -> None:
@@ -34,3 +42,23 @@ def test_bin_year_of_construction_as_in_census() -> None:
     output: pd.DataFrame = _bin_year_of_construction_as_in_census(ber)
 
     assert_frame_equal(output, expected_output)
+
+
+def test_transform_ber(ref: ReferenceTest) -> None:
+    """Transformed ber matches reference data.
+
+    Args:
+        ref (ReferenceTest): a tdda plugin used to verify a DataFrame against a file.
+    """
+    ber_raw = pd.read_parquet(BER_RAW)
+
+    """ DOES:
+        - Extract columns: Year of Construction, Dwelling Type, Energy ...
+        - Extract Dublin rows
+        - Map Period Built data to National Regulatory Periods
+        - Calculate Total Heat Demand
+        - Extract columns: Period Built, Dwelling Type, Total Heat Demand
+    """
+    ber_clean = drem.transform_ber.run(ber_raw)
+
+    ref.assertDataFrameCorrect(ber_clean, BER_CLEAN)
